+++ skins/weewx-wdc/src/js/diagrams/calendar.tsx
@@ -2,6 +2,7 @@ import React, { FunctionComponent } from "react";
 import { ResponsiveCalendar } from "@nivo/calendar";
 import moment from "moment";
 import { NumberValue, scaleQuantize } from "d3-scale";
+import { useMediaQuery } from "@react-hook/media-query";
 
 import { ColorScale } from "@nivo/calendar";
 
@@ -11,21 +12,31 @@ import { TooltipCalendar } from "../components/tooltip-calendar";
 export const CalendarDiagram: FunctionComponent<CalendarDiagramBaseProps> = (
   props: CalendarDiagramBaseProps
 ): React.ReactElement => {
+  const small = useMediaQuery("(max-width: 672px)");
   const start = props.data[0].day;
   const end = props.data[props.data.length - 1].day;
   const yearDiff = moment(end).diff(start, "years");
 
   // @see https://nivo.rocks/storybook/?path=/story/calendarcanvas--custom-color-space-function
+  // @see https://github.com/plouc/nivo/issues/744#issuecomment-573340879
   const color = (): ColorScale => {
+    // @todo use https://github.com/d3/d3-scale#threshold-scales
     const defaultColorScale = scaleQuantize<string>()
       .domain([0, Math.max(...props.data.map((item) => item.value))])
       .range(props.color);
+
     const colorScale = (value: NumberValue) => {
-      return defaultColorScale(value); //adding alpha channel
+      const number = parseFloat(String(value).replace(props.unit, "").trim());
+      return defaultColorScale(number); //adding alpha channel
     };
     colorScale.ticks = (count: number | undefined) => {
-      return defaultColorScale.ticks(count);
+      return defaultColorScale.ticks(count).map((e) => {
+        console.log("e", e);
+        return `${e}${props.unit}`;
+      });
     };
+
+    /*tslint-ignore*/
     return colorScale;
   };
 
@@ -35,7 +46,7 @@ export const CalendarDiagram: FunctionComponent<CalendarDiagramBaseProps> = (
       <div
         className="calendar-diagram"
         // @todo Add responsive style.
-        style={{ height: `${(yearDiff + 1) * 15}vw` }}
+        style={{ height: `${(yearDiff + 1) * (small ? 20 : 12)}vw` }}
       >
         <ResponsiveCalendar
           from={props.data[0].day}
@@ -44,7 +55,12 @@ export const CalendarDiagram: FunctionComponent<CalendarDiagramBaseProps> = (
           emptyColor="#e0e0e0"
           //colors={props.color}
           colorScale={color()}
-          margin={{ top: 40, right: 20, bottom: 40, left: 25 }}
+          margin={{
+            top: 20,
+            right: small ? 20 : 100,
+            bottom: small ? 40 : 20,
+            left: 25,
+          }}
           dayBorderColor="#ffffff"
           monthSpacing={2}
           monthBorderColor="#ffffff"
@@ -60,6 +76,19 @@ export const CalendarDiagram: FunctionComponent<CalendarDiagramBaseProps> = (
               color={props.color[0]}
             />
           )}
+          legends={[
+            {
+              anchor: small ? "bottom-left" : "right",
+              direction: small ? "row" : "column",
+              translateY: small ? -20 : 0,
+              translateX: small ? 0 : -20,
+              itemCount: 4,
+              itemWidth: small ? 50 : 80,
+              itemHeight: 20,
+              itemsSpacing: small ? 2 : 5,
+              itemDirection: "left-to-right",
+            },
+          ]}
         />
       </div>
     </>