+++ changelog.md
@@ -69,4 +69,4 @@
 
 - Added test_install.sh
 - Added "Climatological Days" Tile to stats and year pages
-- Added rain days calendar diagram
+- Added rain days and average temperature calendar diagram